pyproject.toml
[project]
name = "vendi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "flask-sock>=0.7.0",
    "google-cloud-vision>=3.13.0",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "piper-tts>=1.3.0",
    "faster-whisper>=1.0.3",
    "transformers>=4.45.0",
    "torch>=2.4.0",
    "python-dotenv>=1.0.1",
    "pillow>=12.2.0",
    "torchvision>=0.26.0",
    "accelerate>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]

# Force torch to come from PyTorch's CUDA wheel index instead of PyPI's CPU build.
[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }
